+++ alphastats/dataset/DataSet.py
@@ -3,12 +3,12 @@ from typing import Dict, List, Optional, Tuple, Union
 import pandas as pd
 import plotly
 import scipy
+from dataset.dataset_factory import DataSetFactory
+from dataset.dataset_harmonizer import DataHarmonizer
+from dataset.DataSet_Plot import Plot
+from dataset.DataSet_Preprocess import Preprocess
+from dataset.DataSet_Statistics import Statistics
 
-from alphastats.dataset_factory import DataSetFactory
-from alphastats.dataset_harmonizer import DataHarmonizer
-from alphastats.DataSet_Plot import Plot
-from alphastats.DataSet_Preprocess import Preprocess
-from alphastats.DataSet_Statistics import Statistics
 from alphastats.keys import Cols
 from alphastats.loader.BaseLoader import BaseLoader
 from alphastats.plots.clustermap import ClusterMap
