+++ alphastats/dataset/dataset_factory.py
@@ -4,8 +4,8 @@ from typing import List, Optional, Tuple, Union
 
 import numpy as np
 import pandas as pd
+from dataset.dataset_harmonizer import DataHarmonizer
 
-from alphastats.dataset_harmonizer import DataHarmonizer
 from alphastats.keys import Cols
