+++ alphastats/gui/pages/02_Import Data.py
@@ -1,6 +1,6 @@
 import streamlit as st
+from dataset.DataSet import DataSet
 
-from alphastats.DataSet import DataSet
 from alphastats.gui.utils.import_helper import (
     load_example_data,
     load_proteomics_data,
