+++ alphastats/gui/utils/analysis.py
@@ -6,8 +6,8 @@ from typing import Any, Dict, Optional, Tuple, Union
 
 import pandas as pd
 import streamlit as st
+from dataset.DataSet import DataSet
 
-from alphastats.DataSet import DataSet
 from alphastats.keys import Cols, ConstantsClass
 from alphastats.plots.plot_utils import PlotlyObject
 from alphastats.plots.volcano_plot import VolcanoPlot
