+++ alphastats/gui/utils/import_helper.py
@@ -5,9 +5,9 @@ from typing import List, Optional, Tuple
 
 import pandas as pd
 import streamlit as st
+from dataset.DataSet import DataSet
 from streamlit.runtime.uploaded_file_manager import UploadedFile
 
-from alphastats.DataSet import DataSet
 from alphastats.gui.utils.options import SOFTWARE_OPTIONS
 from alphastats.loader.maxquant_loader import BaseLoader, MaxQuantLoader
