+++ alphastats/gui/utils/preprocessing_helper.py
@@ -2,9 +2,9 @@ from typing import List
 
 import pandas as pd
 import streamlit as st
+from dataset.DataSet import DataSet
 from st_cytoscape import cytoscape
 
-from alphastats.DataSet import DataSet
 from alphastats.keys import Cols
 
 CYTOSCAPE_STYLESHEET = [
