+++ alphastats/llm/llm_functions.py
@@ -3,8 +3,8 @@
 from typing import Dict, List
 
 import pandas as pd
+from dataset.DataSet import DataSet
 
-from alphastats.DataSet import DataSet
 from alphastats.llm.enrichment_analysis import get_enrichment_data
 from alphastats.llm.uniprot_utils import get_gene_function
