+++ alphastats/llm/llm_integration.py
@@ -6,11 +6,11 @@ from typing import Any, Dict, List, Optional, Tuple
 
 import pandas as pd
 import plotly.io as pio
+from dataset.DataSet import DataSet
 from IPython.display import HTML, Markdown, display
 from openai import OpenAI
 from openai.types.chat import ChatCompletion, ChatCompletionMessageToolCall
 
-from alphastats.DataSet import DataSet
 from alphastats.llm.llm_functions import (
     GENERAL_FUNCTION_MAPPING,
     get_assistant_functions,
