+++ alphastats/llm/prompts.py
@@ -3,9 +3,9 @@
 import os
 from typing import Any, Dict, List
 
+from dataset.DataSet import DataSet
 from openai.types.chat import ChatCompletionMessageToolCall
 
-from alphastats.DataSet import DataSet
 from alphastats.llm.llm_utils import get_subgroups_for_each_group
