+++ alphastats/plots/clustermap.py
@@ -3,8 +3,8 @@ from typing import Dict
 
 import pandas as pd
 import seaborn as sns
+from dataset.DataSet_Statistics import Statistics
 
-from alphastats.DataSet_Statistics import Statistics
 from alphastats.keys import Cols
 from alphastats.plots.plot_utils import PlotUtils
