+++ alphastats/plots/volcano_plot.py
@@ -5,9 +5,9 @@ import pandas as pd
 import plotly
 import plotly.express as px
 import plotly.graph_objects as go
+from dataset.DataSet_Preprocess import PreprocessingStateKeys
+from dataset.DataSet_Statistics import Statistics
 
-from alphastats.DataSet_Preprocess import PreprocessingStateKeys
-from alphastats.DataSet_Statistics import Statistics
 from alphastats.keys import Cols
 from alphastats.multicova import multicova
 from alphastats.plots.plot_utils import PlotlyObject, PlotUtils
